refactor: colocate hook test suites and add changeset API examples
Move the createOnScorerHook suite next to mastra/hooks.ts (plus a new
terminal-run guard test) and the runScorer sampling decision suite into
evals/hooks.test.ts. Add public API examples to the remaining scoring
changesets per review feedback.
